% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.smoots}
\alias{fitted.smoots}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{smoots}(object, ...)
}
\arguments{
\item{object}{an object from the \code{smoots} class.}

\item{...}{included for consistency with the generic function.}
}
\value{
Fitted values extracted from a \code{smoots} class object.
}
\description{
Generic function which extracts fitted values from a \code{smoots} class
object. Both \code{fitted} and \code{fitted.values} can be called.
}
\author{
\itemize{
\item Sebastian Letmathe (Scientific Employee) (Department of Economics,
Paderborn University), \cr
}
}
