% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.smoots.R
\name{plot.smoots}
\alias{plot.smoots}
\title{Plot Method for the Package 'smoots'}
\usage{
\method{plot}{smoots}(x, t = NULL, rescale = TRUE, which = NULL, ...)
}
\arguments{
\item{x}{an input object of class \code{smoots}.}

\item{t}{an optional vector with time points that will be considered for
the x-axis within the plot; is set to NULL by default and uses a vector
\code{1:length(x$ye)} for time points.}

\item{rescale}{a single logical value; is set to \code{TRUE} by default;
if the output of a derivative estimation process is passed to \code{x} and if
\code{rescale = TRUE}, the estimates will be rescaled according to \code{t}.}

\item{which}{a selector for the plot type so that the interactive prompt is
avoided; for the default, \code{which = NULL}, the user will be asked
interactively via the console which plot to show; to avoid this behavior,
set \code{which} to the corresponding number of the plot you would like
to create (1: original series, 2: trend series, 3: residual series,
4: original series with trend series for trend estimation objects,
1: original series, 2: derivative series for trend derivative estimation
object).}

\item{...}{additional arguments of the standard plot method.}
}
\value{
None
}
\description{
This function regulates how objects created by the package \code{smoots} are
plotted.
}
\author{
\itemize{
\item Dominik Schulz (Research Assistant) (Department of Economics, Paderborn
University), \cr
Package Creator and Maintainer
}
}
