% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.smoots.R
\name{print.smoots}
\alias{print.smoots}
\title{Print Method for the Package 'smoots'}
\usage{
\method{print}{smoots}(x, ...)
}
\arguments{
\item{x}{an input object of class \emph{smoots}.}

\item{...}{additional arguments of the standard print method}
}
\value{
None
}
\description{
This function regulates how objects created by the package smoots are printed.
}
\author{
\itemize{
\item Dominik Schulz (Student Assistant) (Department of Economics, Paderborn
University), \cr
Package Creator and Maintainer
}
}
