% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.smoots}
\alias{residuals.smoots}
\title{Extract Model Residuals}
\usage{
\method{residuals}{smoots}(object, ...)
}
\arguments{
\item{object}{an object from the \code{smoots} class.}

\item{...}{included for consistency with the generic function.}
}
\value{
Residuals extracted from a \code{smoots} class object.
}
\description{
Generic function which extracts model residuals from a \code{smoots} class
object. Both \code{residuals} and its abbreviated form \code{resid} can be called.
}
\author{
\itemize{
\item Sebastian Letmathe (Scientific Employee) (Department of Economics,
Paderborn University), \cr
}
}
