% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vix.R
\docType{data}
\name{vix}
\alias{vix}
\title{CBOE Volatility Index (VIX) Financial Time Series Data}
\format{
A data frame with 7452 rows and 9 variables:
\describe{
  \item{Year}{the observation year}
  \item{Month}{the observation month}
  \item{Day}{the observation day}
  \item{Open}{the opening price of the day}
  \item{High}{the highest price of the day}
  \item{Low}{the lowest price of the day}
  \item{Close}{the closing price of the day}
  \item{AdjClose}{the adjusted closing price of the day}
  \item{Volume}{the traded volume}
}
}
\source{
The data was obtained from Yahoo Finance.

\url{https://query1.finance.yahoo.com/v7/finance/download/^VIX?period1=631148400&period2=1564524000&interval=1d&events=history&crumb=Iaq1EPZAQRb}
}
\usage{
vix
}
\description{
A dataset that contains the daily financial data of the VIX from
1990 to July 2019 (currency in USD).
}
\keyword{datasets}
