% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.smoots.R
\name{plot.smoots}
\alias{plot.smoots}
\title{Plot Method for the Package 'smoots'}
\usage{
\method{plot}{smoots}(x, type = "l", main = NULL, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{an input object of class \emph{smoots}.}

\item{type}{the usual 'type' argument of the plot function; is set to "l" per
default.}

\item{main}{the usual 'main' argument of the plot function; is set to NULL per
default and then uses a predefined title that depends on the chosen smoots
plot.}

\item{xlab}{the usual 'xlab' argument of the plot function; is set to NULL
per default and then uses a predefined label for the x-axis that depends on
the chosen smoots plot.}

\item{ylab}{the usual 'ylab' argument of the plot function; is set to NULL
per default and then uses a predefined label for the y-axis that depends on
the chosen smoots plot.}

\item{...}{additional arguments of the standard plot method}
}
\value{
None
}
\description{
This function regulates how objects created by the package smoots are plotted.
}
\author{
\itemize{
\item Dominik Schulz (Student Assistant) (Department of Economics, Paderborn
University), \cr
Package Creator and Maintainer
}
}
