% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempNH.R
\docType{data}
\name{tempNH}
\alias{tempNH}
\title{Mean Monthly Northern Hemisphere Temperature Changes}
\format{A data frame with 1668 rows and 3 variables:
\describe{
  \item{Year}{the observation year}
  \item{Month}{the observation month}
  \item{Change}{the observed mean monthly temperature changes in degrees
    Celsius}
}}
\source{
The data was obtained from the Goddard Institute for Space Studies
(part of the National Aeronautics and Space Administration [NASA]).

\url{https://data.giss.nasa.gov/gistemp/tabledata_v4/NH.Ts+dSST.txt}
}
\usage{
tempNH
}
\description{
A dataset that contains seasonally adjusted mean monthly Northern Hemisphere temperature
changes from 1880 to 2018.
}
\keyword{datasets}
